fix: redirect from api route
