+++ domains/auth/types/auth.ts
@@ -1,0 +1,7 @@
+export interface AuthDataType {
+  accessToken: string;
+  refreshToken: string;
+  refreshTokenExpiresIn: number;
+  expiresIn: number;
+  firstLogin: boolean;
+}
